Hide on scrolling, fix https://github.com/nextcloud/android/issues/6238
