+++ src/main/res/layout/toolbar_standard.xml
@@ -30,7 +30,8 @@
     <RelativeLayout
         android:id="@+id/default_toolbar"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        app:layout_scrollFlags="scroll|enterAlways">
 
         <FrameLayout
             android:id="@+id/preview_image_frame"
@@ -84,7 +85,8 @@
         app:cardCornerRadius="8dp"
         app:cardElevation="2dp"
         app:strokeWidth="0dp"
-        tools:visibility="visible">
+        tools:visibility="visible"
+        app:layout_scrollFlags="scroll|enterAlways">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
@@ -136,6 +138,49 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
 
     </com.google.android.material.card.MaterialCardView>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/sort_list_button_group"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:visibility="gone"
+        app:layout_scrollFlags="scroll|enterAlways">
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/sort_button"
+            style="@style/Widget.MaterialComponents.Button.TextButton"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="8dp"
+            android:contentDescription="@string/actionbar_sort"
+            android:minWidth="0dp"
+            android:text="@string/menu_item_sort_by_date_newest_first"
+            android:textAlignment="textStart"
+            android:textAllCaps="false"
+            android:textColor="@color/fontAppbar"
+            android:textSize="14sp"
+            app:icon="@drawable/ic_keyboard_arrow_down"
+            app:iconGravity="textEnd"
+            app:iconSize="16dp"
+            app:iconTint="@color/fontAppbar"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/switch_grid_view_button"
+            style="@style/Widget.AppTheme.Button.IconButton"
+            android:layout_width="38dp"
+            android:layout_height="38dp"
+            android:layout_marginEnd="4dp"
+            android:contentDescription="@string/action_switch_grid_view"
+            app:cornerRadius="24dp"
+            app:icon="@drawable/ic_view_module"
+            app:iconTint="@color/fontAppbar"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
     <include layout="@layout/info_box" />
